refactor: create a conceptual-process service that will do the util method
- Removed the util as we need the store, don't like it passing it on
- Use the service in both places so they do the same logic

Refs: oph-716
